examples/pdf_to_images_agent/pyproject.toml
[project]
name = "pdf-to-images-agent"
version = "0.1.0"
description = "PDF to page images conversion agent example"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.3.18",
    "langchain-core>=0.3.34",
    "langchain-openai>=0.3.5",
    "middleman-ai",
    "python-dotenv>=1.0.1",
]

[tool.uv]
[tool.uv.sources]
middleman-ai = { path = "../../", editable = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = [
    "src/pdf_to_images_agent",
]
